LayoutTests/platform/chromium-win/fast/dom/Element/getBoundingClientRect-expected.txt

Client bounding rect for #1
testRect.top 44
testRect.left 8
testRect.width 300
testRect.height 100
testRect.bottom 144
testRect.right 308

FAIL testRect.top should be 42. Was 44.
PASS testRect.left is 8
PASS testRect.width is 300
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #2
testRect.top 196
testRect.left 8
testRect.width 320
testRect.height 120
testRect.bottom 316
testRect.right 328

FAIL testRect.top should be 192. Was 196.
PASS testRect.left is 8
PASS testRect.width is 320
PASS testRect.height is 120
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #3
testRect.top 368
testRect.left 18
testRect.width 300
testRect.height 100
testRect.bottom 468
testRect.right 318

FAIL testRect.top should be 362. Was 368.
PASS testRect.left is 18
PASS testRect.width is 300
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #4
testRect.top 499
testRect.left 37
testRect.width 142
testRect.height 142
testRect.bottom 641
testRect.right 179

FAIL testRect.top should be 491. Was 499.
PASS testRect.left is 37
PASS testRect.width is 142
PASS testRect.height is 142
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #5
testRect.top 672
testRect.left 8
testRect.width 300
testRect.height 261
testRect.bottom 933
testRect.right 308

FAIL testRect.top should be 662. Was 672.
PASS testRect.left is 8
PASS testRect.width is 300
FAIL testRect.height should be 240. Was 261.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #6
Known failure. Skipping.

Client bounding rect for #7
testRect.top 1305
testRect.left 8
testRect.width 295
testRect.height 199
testRect.bottom 1504
testRect.right 303

FAIL testRect.top should be 1236. Was 1305.
PASS testRect.left is 8
FAIL testRect.width should be 299. Was 295.
FAIL testRect.height should be 180. Was 199.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #8
testRect.top 1557
testRect.left 8
testRect.width 300
testRect.height 50
testRect.bottom 1607
testRect.right 308

FAIL testRect.top should be 1466. Was 1557.
PASS testRect.left is 8
PASS testRect.width is 300
FAIL testRect.height should be 46. Was 50.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #9
testRect.top 1659
testRect.left 8
testRect.width 300
testRect.height 70
testRect.bottom 1729
testRect.right 308

FAIL testRect.top should be 1562. Was 1659.
PASS testRect.left is 8
PASS testRect.width is 300
FAIL testRect.height should be 64. Was 70.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #10
testRect.top 1783
testRect.left 8
testRect.width 300
testRect.height 22
testRect.bottom 1805
testRect.right 308

FAIL testRect.top should be 1678. Was 1783.
PASS testRect.left is 8
PASS testRect.width is 300
FAIL testRect.height should be 20. Was 22.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #11
testRect.top 1885
testRect.left 10
testRect.width 147
testRect.height 22
testRect.bottom 1907
testRect.right 157

FAIL testRect.top should be 1774. Was 1885.
PASS testRect.left is 10
PASS testRect.width is 147
FAIL testRect.height should be 20. Was 22.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #12
testRect.top 1985
testRect.left 8
testRect.width 100
testRect.height 100
testRect.bottom 2085
testRect.right 108

FAIL testRect.top should be 1868. Was 1985.
PASS testRect.left is 8
PASS testRect.width is 100
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #13
testRect.top 2222
testRect.left 8
testRect.width 175
testRect.height 19
testRect.bottom 2241
testRect.right 183

FAIL testRect.top should be 2104. Was 2222.
PASS testRect.left is 8
FAIL testRect.width should be 180. Was 175.
FAIL testRect.height should be 18. Was 19.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #14
testRect.top 2294
testRect.left 8
testRect.width 300
testRect.height 239
testRect.bottom 2533
testRect.right 308

FAIL testRect.top should be 2172. Was 2294.
PASS testRect.left is 8
FAIL testRect.width should be 200. Was 300.
FAIL testRect.height should be 236. Was 239.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #15
testRect.top 2586
testRect.left 8
testRect.width 0
testRect.height 0
testRect.bottom 2586
testRect.right 8

FAIL testRect.top should be 2458. Was 2586.
PASS testRect.left is 8
PASS testRect.width is 0
PASS testRect.height is 0
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

PASS successfullyParsed is true

TEST COMPLETE
